+++ affinda/models/_models.py
@@ -521,6 +521,8 @@ class Collection(msrest.serialization.Model):
     :vartype auto_validation_threshold: float
     :ivar fields:
     :vartype fields: list[~affinda.models.FieldGroup]
+    :ivar fields_layout:
+    :vartype fields_layout: ~affinda.models.FieldsLayout
     :ivar fields_configured:
     :vartype fields_configured: bool
     :ivar date_format_preference: Known values are: "DMY", "MDY", "YMD".
@@ -550,6 +552,7 @@ class Collection(msrest.serialization.Model):
         "extractor": {"key": "extractor", "type": "Extractor"},
         "auto_validation_threshold": {"key": "autoValidationThreshold", "type": "float"},
         "fields": {"key": "fields", "type": "[FieldGroup]"},
+        "fields_layout": {"key": "fieldsLayout", "type": "FieldsLayout"},
         "fields_configured": {"key": "fieldsConfigured", "type": "bool"},
         "date_format_preference": {"key": "dateFormatPreference", "type": "str"},
         "date_format_from_document": {"key": "dateFormatFromDocument", "type": "bool"},
@@ -573,6 +576,8 @@ class Collection(msrest.serialization.Model):
         :paramtype auto_validation_threshold: float
         :keyword fields:
         :paramtype fields: list[~affinda.models.FieldGroup]
+        :keyword fields_layout:
+        :paramtype fields_layout: ~affinda.models.FieldsLayout
         :keyword fields_configured:
         :paramtype fields_configured: bool
         :keyword date_format_preference: Known values are: "DMY", "MDY", "YMD".
@@ -597,6 +602,7 @@ class Collection(msrest.serialization.Model):
         self.extractor = kwargs.get("extractor", None)
         self.auto_validation_threshold = kwargs.get("auto_validation_threshold", None)
         self.fields = kwargs.get("fields", None)
+        self.fields_layout = kwargs.get("fields_layout", None)
         self.fields_configured = kwargs.get("fields_configured", None)
         self.date_format_preference = kwargs.get("date_format_preference", None)
         self.date_format_from_document = kwargs.get("date_format_from_document", None)
@@ -3887,6 +3893,97 @@ class Field(msrest.serialization.Model):
 
     :ivar label: Required.
     :vartype label: str
+    :ivar data_point: Required. Data point identifier.
+    :vartype data_point: str
+    :ivar mandatory:
+    :vartype mandatory: bool
+    :ivar auto_validation_threshold:
+    :vartype auto_validation_threshold: float
+    :ivar show_dropdown:
+    :vartype show_dropdown: bool
+    :ivar fields:
+    :vartype fields: list[~affinda.models.Field]
+    """
+
+    _validation = {
+        "label": {"required": True},
+        "data_point": {"required": True},
+        "auto_validation_threshold": {"maximum": 1, "minimum": 0.9},
+    }
+
+    _attribute_map = {
+        "label": {"key": "label", "type": "str"},
+        "data_point": {"key": "dataPoint", "type": "str"},
+        "mandatory": {"key": "mandatory", "type": "bool"},
+        "auto_validation_threshold": {"key": "autoValidationThreshold", "type": "float"},
+        "show_dropdown": {"key": "showDropdown", "type": "bool"},
+        "fields": {"key": "fields", "type": "[Field]"},
+    }
+
+    def __init__(self, **kwargs):
+        """
+        :keyword label: Required.
+        :paramtype label: str
+        :keyword data_point: Required. Data point identifier.
+        :paramtype data_point: str
+        :keyword mandatory:
+        :paramtype mandatory: bool
+        :keyword auto_validation_threshold:
+        :paramtype auto_validation_threshold: float
+        :keyword show_dropdown:
+        :paramtype show_dropdown: bool
+        :keyword fields:
+        :paramtype fields: list[~affinda.models.Field]
+        """
+        super(Field, self).__init__(**kwargs)
+        self.label = kwargs["label"]
+        self.data_point = kwargs["data_point"]
+        self.mandatory = kwargs.get("mandatory", None)
+        self.auto_validation_threshold = kwargs.get("auto_validation_threshold", None)
+        self.show_dropdown = kwargs.get("show_dropdown", None)
+        self.fields = kwargs.get("fields", None)
+
+
+class FieldCategory(msrest.serialization.Model):
+    """FieldCategory.
+
+    All required parameters must be populated in order to send to Azure.
+
+    :ivar enabled_fields: Required.
+    :vartype enabled_fields: list[~affinda.models.Field]
+    :ivar disabled_fields: Required.
+    :vartype disabled_fields: list[~affinda.models.Field]
+    """
+
+    _validation = {
+        "enabled_fields": {"required": True},
+        "disabled_fields": {"required": True},
+    }
+
+    _attribute_map = {
+        "enabled_fields": {"key": "enabledFields", "type": "[Field]"},
+        "disabled_fields": {"key": "disabledFields", "type": "[Field]"},
+    }
+
+    def __init__(self, **kwargs):
+        """
+        :keyword enabled_fields: Required.
+        :paramtype enabled_fields: list[~affinda.models.Field]
+        :keyword disabled_fields: Required.
+        :paramtype disabled_fields: list[~affinda.models.Field]
+        """
+        super(FieldCategory, self).__init__(**kwargs)
+        self.enabled_fields = kwargs["enabled_fields"]
+        self.disabled_fields = kwargs["disabled_fields"]
+
+
+class FieldDeprecated(msrest.serialization.Model):
+    """FieldDeprecated.
+
+    All required parameters must be populated in order to send to Azure.
+
+    :ivar label: Required.
+    :vartype label: str
     :ivar slug:
     :vartype slug: str
     :ivar data_point: Required.
@@ -3900,7 +3997,7 @@ class Field(msrest.serialization.Model):
     :ivar show_dropdown:
     :vartype show_dropdown: bool
     :ivar fields:
-    :vartype fields: list[~affinda.models.Field]
+    :vartype fields: list[~affinda.models.FieldDeprecated]
     """
 
     _validation = {
@@ -3916,7 +4013,7 @@ class Field(msrest.serialization.Model):
         "disabled": {"key": "disabled", "type": "bool"},
         "auto_validation_threshold": {"key": "autoValidationThreshold", "type": "float"},
         "show_dropdown": {"key": "showDropdown", "type": "bool"},
-        "fields": {"key": "fields", "type": "[Field]"},
+        "fields": {"key": "fields", "type": "[FieldDeprecated]"},
     }
 
     def __init__(self, **kwargs):
@@ -3936,9 +4033,9 @@ class Field(msrest.serialization.Model):
         :keyword show_dropdown:
         :paramtype show_dropdown: bool
         :keyword fields:
-        :paramtype fields: list[~affinda.models.Field]
+        :paramtype fields: list[~affinda.models.FieldDeprecated]
         """
-        super(Field, self).__init__(**kwargs)
+        super(FieldDeprecated, self).__init__(**kwargs)
         self.label = kwargs["label"]
         self.slug = kwargs.get("slug", None)
         self.data_point = kwargs["data_point"]
@@ -3957,7 +4054,7 @@ class FieldGroup(msrest.serialization.Model):
     :ivar label: Required.
     :vartype label: str
     :ivar fields: Required.
-    :vartype fields: list[~affinda.models.Field]
+    :vartype fields: list[~affinda.models.FieldDeprecated]
     """
 
     _validation = {
@@ -3967,7 +4064,7 @@ class FieldGroup(msrest.serialization.Model):
 
     _attribute_map = {
         "label": {"key": "label", "type": "str"},
-        "fields": {"key": "fields", "type": "[Field]"},
+        "fields": {"key": "fields", "type": "[FieldDeprecated]"},
     }
 
     def __init__(self, **kwargs):
@@ -3975,11 +4072,44 @@ class FieldGroup(msrest.serialization.Model):
         :keyword label: Required.
         :paramtype label: str
         :keyword fields: Required.
-        :paramtype fields: list[~affinda.models.Field]
+        :paramtype fields: list[~affinda.models.FieldDeprecated]
         """
         super(FieldGroup, self).__init__(**kwargs)
         self.label = kwargs["label"]
         self.fields = kwargs["fields"]
+
+
+class FieldsLayout(msrest.serialization.Model):
+    """FieldsLayout.
+
+    All required parameters must be populated in order to send to Azure.
+
+    :ivar default_category: Required.
+    :vartype default_category: ~affinda.models.FieldCategory
+    :ivar categories: Required.
+    :vartype categories: list[~affinda.models.FieldCategory]
+    """
+
+    _validation = {
+        "default_category": {"required": True},
+        "categories": {"required": True},
+    }
+
+    _attribute_map = {
+        "default_category": {"key": "defaultCategory", "type": "FieldCategory"},
+        "categories": {"key": "categories", "type": "[FieldCategory]"},
+    }
+
+    def __init__(self, **kwargs):
+        """
+        :keyword default_category: Required.
+        :paramtype default_category: ~affinda.models.FieldCategory
+        :keyword categories: Required.
+        :paramtype categories: list[~affinda.models.FieldCategory]
+        """
+        super(FieldsLayout, self).__init__(**kwargs)
+        self.default_category = kwargs["default_category"]
+        self.categories = kwargs["categories"]
 
 
 class FloatAnnotation(Annotation):
@@ -9274,9 +9404,6 @@ class JobDescriptionData(msrest.serialization.Model):
     :vartype certifications: list[~affinda.models.TextAnnotation]
     :ivar years_experience:
     :vartype years_experience: ~affinda.models.YearsExperienceAnnotation
-    :ivar raw_text: All of the raw text of the parsed job description, example is shortened for
-     readability.
-    :vartype raw_text: str
     """
 
     _attribute_map = {
@@ -9301,7 +9428,6 @@ class JobDescriptionData(msrest.serialization.Model):
         "location": {"key": "location", "type": "LocationAnnotation"},
         "certifications": {"key": "certifications", "type": "[TextAnnotation]"},
         "years_experience": {"key": "yearsExperience", "type": "YearsExperienceAnnotation"},
-        "raw_text": {"key": "rawText", "type": "str"},
     }
 
     def __init__(self, **kwargs):
@@ -9343,9 +9469,6 @@ class JobDescriptionData(msrest.serialization.Model):
         :paramtype certifications: list[~affinda.models.TextAnnotation]
         :keyword years_experience:
         :paramtype years_experience: ~affinda.models.YearsExperienceAnnotation
-        :keyword raw_text: All of the raw text of the parsed job description, example is shortened for
-         readability.
-        :paramtype raw_text: str
         """
         super(JobDescriptionData, self).__init__(**kwargs)
         self.additional_properties = kwargs.get("additional_properties", None)
@@ -9366,7 +9489,6 @@ class JobDescriptionData(msrest.serialization.Model):
         self.location = kwargs.get("location", None)
         self.certifications = kwargs.get("certifications", None)
         self.years_experience = kwargs.get("years_experience", None)
-        self.raw_text = kwargs.get("raw_text", None)
 
 
 class JobDescriptionSearch(msrest.serialization.Model):